+++ R/normaliseCounts.R
@@ -13,7 +13,8 @@
 #'
 #' @examples
 #' data(test.dge)
-#' normaliseCounts(test.dge, method = "CPM", threshold = 0)
+#' test.dge.normalized <- normaliseCounts(test.dge, method = "CPM", threshold = 0)
+#' head(test.dge.normalized)
 #'
 
 normaliseCounts <-