+++ NEWS.md
@@ -1,3 +1,11 @@
+# bartools 1.1.0
+
+- Fix plotDetectedBarcodes and calcPercentileBarcodes to be inclusive, i.e. at least 1 barcode makes up top x percentile of a sample
+- improve coloring of barcodes in bubble plots
+- update plotDetectedBarcodes
+- collapseReplicates and thresholdCounts update the lib.size column to the sum of the merged replicates or filtered sample.
+- fix bug where small dots were plotted for barcodes that were not present in sample in plotOrderedBubble and plotBarcodeBubble
+
 # bartools 1.0.0
 
 - Updated function parameter inputs across whole package